+++ pkg/interfaces/serial_race_test.go
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
 // Copyright (c) 2024-2026 Quad4.io
 
-//go:build (linux || darwin || freebsd || openbsd || windows) && !js
+//go:build !js
 
 package interfaces